backend/pyproject.toml
[project]
name = "parrotsoftware-treatment-api"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.14"
dependencies = [
    "fastapi>=0.127.0",
    "google-cloud-speech>=2.35.0",
    "google-cloud-texttospeech>=2.33.0",
    "httpx>=0.28.1",
    "openai>=2.14.0",
    "pydantic>=2.12.5",
    "pydantic-settings>=2.12.0",
    "pyjwt>=2.10.1",
    "python-dotenv>=1.2.1",
    "resend>=2.19.0",
    "uvicorn[standard]>=0.40.0",
]
